requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.1.3
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pypdf==5.1.0
